refactor: extract ACP turn runner
